Improve code to clean browsing data
