+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "flowr-desktop",
-  "version": "0.3.6-delta",
+  "version": "0.3.7",
   "author": {
     "email": "hj@taktik.com",
     "name": "Hubert Jacquin"
