+++ src/launcher/index.ts
@@ -1,10 +1,9 @@
 import { ipcMain, app, BrowserWindow } from 'electron'
 import { resolve } from 'path'
 import { homedir } from 'os'
-import { removeSync, remove } from 'fs-extra'
-import { autoUpdater } from 'electron-updater'
 import { createFlowrWindow, initFlowrConfig, buildBrowserWindowConfig, FRONTEND_CONFIG_NAME, DEFAULT_FRONTEND_STORE } from '../frontend'
 import { createWexondWindow, setWexondLog } from '~/main'
+import { clearBrowsingData } from '~/main/clearBrowsingData'
 import { getMigrateUserPreferences } from './migration/fromFlowrClientToFlowrPcClient'
 import { FlowrWindow } from 'src/frontend/flowr-window'
 export const log = require('electron-log')
@@ -24,7 +23,7 @@ app.commandLine.appendSwitch('widevine-cdm-path', resolve('/Applications/Google 
 // The version of plugin can be got from `chrome://components` page in Chrome.
 app.commandLine.appendSwitch('widevine-cdm-version', '4.10.1303.2')
 const userAppData = resolve(homedir(), '.flowr-electron')
-removeSync(userAppData)
+
 app.setPath('userData', userAppData)
 log.transports.file.level = 'verbose'
 log.transports.file.file = resolve(app.getPath('userData'), 'log.log')
@@ -54,30 +53,13 @@ process.on('uncaughtException', error => {
 })
 
 app.on('ready', async () => {
-
+  clearBrowsingData()
   app.on('activate', async () => {
     if (flowrWindow === null) {
       await initFlowr()
     }
   })
   await initFlowr()
-
-  autoUpdater.on('update-downloaded', ({ version }) => {
-    // TODO
-    if (flowrWindow) {
-      flowrWindow.webContents.send('update-available', version)
-    }
-  })
-
-  ipcMain.on('update-install', () => {
-    autoUpdater.quitAndInstall()
-  })
-
-  ipcMain.on('update-check', () => {
-    if (process.env.ENV !== 'dev') {
-      autoUpdater.checkForUpdates()
-    }
-  })
 
   ipcMain.on('window-focus', () => {
     if (flowrWindow) {
@@ -118,11 +100,7 @@ app.on('ready', async () => {
   })
 })
 
-ipcMain.on('clear-application-data', async () => {
-  await remove(app.getPath('userData'))
-  app.relaunch()
-  app.exit()
-})
+ipcMain.on('clear-application-data', clearBrowsingData)
 
 app.on('window-all-closed', () => {
   applicationManager.destroy()
